Added      app/Areas/User/Views/ActionLog/Index.sword Added      app/Areas/User/Views/ActionLog/Latest.sword Added      app/Areas/User/Views/LoginLog/Index.sword Added      app/Areas/User/Views/LoginLog/Latest.sword Added      app/Controllers/TestController.php Added      devops/deploy.sh Modified   .env.example Modified   README.md Modified   app/Application.php Modified   app/Areas/Menu/Controllers/GroupController.php Modified   app/Areas/Menu/Controllers/ItemController.php Modified   app/Areas/Menu/Views/Group.sword Modified   app/Areas/Menu/Views/Item.sword Missing    app/Areas/Rbac/Components/Rbac.php Modified   app/Areas/Rbac/Controllers/AdminController.php Modified   app/Areas/Rbac/Controllers/AdminRoleController.php Modified   app/Areas/Rbac/Controllers/IndexController.php Modified   app/Areas/Rbac/Controllers/PermissionController.php Modified   app/Areas/Rbac/Controllers/RoleController.php Modified   app/Areas/Rbac/Controllers/RolePermissionController.php Modified   app/Areas/Rbac/Models/Role.php Modified   app/Areas/Rbac/Views/Admin.sword Modified   app/Areas/Rbac/Views/AdminRole.sword Modified   app/Areas/Rbac/Views/Permission.sword Modified   app/Areas/Rbac/Views/Role.sword Modified   app/Areas/Rbac/Views/RolePermission.sword Modified   app/Areas/User/Controllers/AccountController.php Modified   app/Areas/User/Controllers/ActionLogController.php Modified   app/Areas/User/Controllers/LoginLogController.php Modified   app/Areas/User/Controllers/PasswordController.php Modified   app/Areas/User/Controllers/SessionController.php Modified   app/Areas/User/Services/ResetPasswordTokenService.php Modified   app/Areas/User/Views/Account/Register.sword Modified   app/Areas/User/Views/Password/Change.sword Modified   app/Areas/User/Views/Password/Forget.sword Modified   app/Areas/User/Views/Password/Reset.sword Modified   app/Areas/User/Views/Session/Login.sword Modified   app/Controllers/IndexController.php Missing    app/Messages/en/Module.php Missing    app/Messages/en/User.php Missing    app/Messages/zh-CN/Module.php Missing    app/Messages/zh-CN/User.php Modified   app/Models/Admin.php Modified   app/Models/AdminActionLog.php Modified   app/Models/AdminLoginLog.php Modified   app/Models/AdminMenu.php Missing    app/Models/ModelBase.php Modified   app/Plugins/AdminActionLogPlugin.php Missing    app/Views/Errors/500.phtml Modified   app/Views/Index/Index.sword Modified   app/Views/Layouts/Default.sword Modified   app/Views/SystemInformation/Index.sword Modified   app/Views/Widgets/SideMenu.sword Modified   app/Widgets/SideMenuWidget.php Modified   config/app.php Modified   docker/docker-compose.yml Modified   docker/fpm/conf.d/fpm.conf Modified   docker/nginx/conf/nginx.conf Modified   docker/swoole.yml Modified   public/.htaccess Modified   public/assets/css/app.css Modified   public/assets/js/app.js Modified   public/index.php Modified   public/static/css/manaphp.css Missing    app/Areas/User/Views/ActionLog.sword Missing    app/Areas/User/Views/LoginLog.sword

+++ app/Areas/User/Controllers/PasswordController.php
@@ -19,51 +19,74 @@ class PasswordController extends Controller
     public function forgetAction()
     {
         if ($this->request->isPost()) {
-            $user_name = input('user_name');
+            $admin_name = input('admin_name');
             $email = input('email');
 
-            $admin = Admin::first(['admin_name' => $user_name]);
+            $admin = Admin::first(['admin_name' => $admin_name]);
             if (!$admin || $admin->email !== $email) {
                 return $this->response->setJsonError('账号不存在或账号与邮箱不匹配');
             }
-            return $this->response->setJsonData(jwt_encode(['user_name' => $user_name, 'scope' => 'admin.user.password.forget'], 300), '重置密码链接已发到您的邮箱');
+
+            $token = jwt_encode(['admin_name' => $admin_name, 'scope' => 'admin.password.forget'], 600);
+
+            $this->mailer->compose()
+                ->setSubject($this->configure->name . '-重置密码邮件')
+                ->setTo($email)
+                ->setHtmlBody(['@app/Areas/User/Views/Mail/ResetPassword', 'email' => $email, 'admin_name' => $admin_name, 'token' => $token])
+                ->send();
+            return $this->response->setJsonOk('重置密码连接已经发送到您的邮箱');
         } else {
             $this->view->setVar('redirect', input('redirect', $this->router->createUrl('/')));
 
-            return $this->view->setVar('user_name', $this->cookies->get('user_name'));
+            return $this->view->setVar('admin_name', $this->cookies->get('admin_name'));
         }
     }
 
     public function resetAction()
     {
         if ($this->request->isAjax()) {
-            $claims = jwt_decode(input('token'), 'admin.user.password.forget');
+            try {
+                $claims = jwt_decode(input('token'), 'admin.password.forget');
+            } catch (\Exception $exception) {
+                return '重置失败：Token已过期';
+            }
 
-            $user_name = $claims['user_name'];
+            $admin_name = $claims['admin_name'];
 
-            $admin = Admin::firstOrFail(['admin_name' => $user_name]);
+            $admin = Admin::firstOrFail(['admin_name' => $admin_name]);
             $admin->password = input('password');
             $admin->update();
 
             return $this->response->setJsonData([], '重置密码成功');
         } else {
-            $claims = jwt_decode(input('token'), 'admin.user.password.forget');
+            $token = input('token');
+
+            try {
+                $claims = jwt_decode($token, 'admin.password.forget');
+            } catch (\Exception $exception) {
+                return $this->view->setVars(['expired' => true]);
+            }
+
+            return $this->view->setVars([
+                'expired' => false,
+                'admin_name' => $claims['admin_name'],
+                'token' => $token,
+            ]);
         }
     }
 
     public function changeAction()
     {
         if ($this->request->isPost()) {
-            if (!$this->configure->debug) {
-                $this->captcha->verify();
-            }
-
             $admin = Admin::get($this->identity->getId());
             if (!$admin->verifyPassword(input('old_password'))) {
                 return '旧密码不正确';
             }
 
             $admin->password = input('new_password');
+            if (input('new_password_confirm') !== $admin->password) {
+                return '两次输入的密码不一致';
+            }
 
             $admin->update();
             $this->session->destroy();
